spider clovis city council: need help w time+date
hey. everything works except _parse_start(), lines#61-76
datetime and regex always gets me.
i can get the time (lines#63-65).
i can get the date string (line#68) but am struggling to extract the substring to only the date portion of the string (line#72) and then combining the date and time together for a singular datetime object. For the time being, using a dummy string for the date on line#69