avenal city council _parse_start()
start time for meetings are noted on the agenda pdf. link to agenda is being extracted in the 'links' data.

at first i was going to make an assumption for meeting time, but looking at the meeting times noted on the pdfs they're so different that i don't think it's safe to assume. 

any advise on extracting info from a pdf? is the repo set up to use PyPDF4 or something like that?